RavenDB-6579 *cached topology now properly written as valid json *get topology moved to databases handler from admin/databasses * Add/Delete databases will properly notify listeners * Cluster tests are now passing
