+++ src/Raven.Server/Web/System/DatabasesHandler.cs
@@ -1,15 +1,19 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Net;
 using System.Threading.Tasks;
 using Raven.Client;
+using Raven.Client.Http;
 using Raven.Client.Server.Operations;
 using Raven.Client.Util;
 using Raven.Server.Config;
 using Raven.Server.Documents;
 using Raven.Server.Extensions;
 using Raven.Server.Json;
+using Raven.Server.Rachis;
 using Raven.Server.Routing;
+using Raven.Server.ServerWide;
 using Raven.Server.ServerWide.Context;
 using Sparrow.Json;
 using Sparrow.Json.Parsing;
@@ -55,6 +59,83 @@ namespace Raven.Server.Web.System
             }
 
             return Task.CompletedTask;
+        }
+
+        [RavenAction("/databases/topology", "GET", "/databases/topology?&name={databaseName:string}&url={url:string}")]
+        public Task GetTopology()
+        {
+            var name = GetQueryStringValueAndAssertIfSingleAndNotEmpty("name");
+            var url = GetStringQueryString("url", false);
+            TransactionOperationContext context;
+            using (ServerStore.ContextPool.AllocateOperationContext(out context))
+            {
+                var dbId = Constants.Documents.Prefix + name;
+                long etag;
+                using (context.OpenReadTransaction())
+                using (var dbBlit = ServerStore.Cluster.Read(context, dbId, out etag))
+                {
+                    if (dbBlit == null)
+                    {
+                        HttpContext.Response.StatusCode = (int)HttpStatusCode.NotFound;
+                        using (var writer = new BlittableJsonTextWriter(context, HttpContext.Response.Body))
+                        {
+                            context.Write(writer,
+                                new DynamicJsonValue
+                                {
+                                    ["Type"] = "Error",
+                                    ["Message"] = "Database " + name + " wasn't found"
+                                });
+                        }
+                        return Task.CompletedTask;
+                    }
+
+                    UnprotectSecuredSettingsOfDatabaseDocument(dbBlit);
+                    var clusterTopology = ServerStore.GetClusterTopology(context);
+                    var dbRecord = JsonDeserializationCluster.DatabaseRecord(dbBlit);
+                    using (var writer = new BlittableJsonTextWriter(context, ResponseBodyStream()))
+                    {
+                        GenerateTopology(context, writer, dbRecord, clusterTopology, etag,url);
+                    }
+                }
+            }
+
+            return Task.CompletedTask;
+        }
+
+        private void GenerateTopology(JsonOperationContext context, BlittableJsonTextWriter writer, DatabaseRecord dbRecord, ClusterTopology clusterTopology, long etag = 0, string url = null)
+        {
+            context.Write(writer, new DynamicJsonValue
+            {
+                [nameof(Topology.LeaderNode)] = new DynamicJsonValue
+                {
+                    //TODO:this should return the senator but for now so it will work I'm returning the "primary" node
+                    [nameof(ServerNode.Url)] = url??Server.Configuration.Core.ServerUrl,
+                    [nameof(ServerNode.Database)] = dbRecord.DatabaseName,
+                },
+                [nameof(Topology.Nodes)] = new DynamicJsonArray(dbRecord.Topology.AllNodes.Select(x => new DynamicJsonValue
+                {
+                    [nameof(ServerNode.Url)] = clusterTopology.GetUrlFormTag(x),
+                    [nameof(ServerNode.Database)] = dbRecord.DatabaseName,
+                })),
+                [nameof(Topology.ReadBehavior)] =
+                ReadBehavior.LeaderWithFailoverWhenRequestTimeSlaThresholdIsReached.ToString(),
+                [nameof(Topology.WriteBehavior)] = WriteBehavior.LeaderOnly.ToString(),
+                [nameof(Topology.SLA)] = new DynamicJsonValue
+                {
+                    [nameof(TopologySla.RequestTimeThresholdInMilliseconds)] = 100,
+                },
+                [nameof(Topology.Etag)] = etag,
+            });
+        }
+
+        private void UnprotectSecuredSettingsOfDatabaseDocument(BlittableJsonReaderObject obj)
+        {
+            //TODO: implement this
+            object securedSettings;
+            if (obj.TryGetMember("SecuredSettings", out securedSettings) == false)
+            {
+
+            }
         }
 
         private Task DbInfo(string dbName)
